requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.14.0
google-cloud-vision>=3.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.28.0
